Create Keyboard events: Keypress.js
The onKeypress event fires when a key is pressed and at an interval determined by the operating system as long as the key is still held down. The continual firing seperates it from the keydown and keyup events, which fire only once.

The event only fires if the target field has focus, though the target can be the entire window to catch the keypress event globally.